type of sample rate changed to double
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@814 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ kwave/MainWidget.cpp
@@ -397,7 +397,7 @@ void MainWidget::refreshChannelControls()
 	min = 0;
 	max = h-m_signal_frame.height();
 	m_scrollbar->setRange(min, max);
-	m_scrollbar->setValue(floor(val * (double)max));
+	m_scrollbar->setValue((int)floor(val * (double)max));
 	m_scrollbar->setSteps(1, m_signal_frame.height());
     }
